// shopsmart-ui/shopsmart-ui/src/app/models/razorpay-models.ts
 
export interface RazorpayOrderRequestDTO {
  orderId: number;
  amount: number; 
  currency: string; 
  receipt: string; 
}

export interface RazorpayOrderResponseDTO {
  id: string; 
  entity: string; 
  amount: number; 
  currency: string;
  receipt: string;
  status: string; 
  attempts: number;
  createdAt: number; 
}

export interface RazorpayPaymentCaptureRequestDTO {
  razorpayPaymentId: string;
  razorpayOrderId: string;
  razorpaySignature: string;
  amount: number;
  internalOrderId: number;
}